// frontend/src/api/types.ts
export interface components {
    schemas: {
        Batch: {
            readonly id: number;
            recipe: number;
            /** Format: decimal */
            batch_size_g: string;
            /** Format: decimal */
            concentration?: string;
            maceration_days?: number;
            /** Format: date */
            made_on: string;
            status?: components["schemas"]["StatusEnum"];
            rating?: number | null;
            /** Format: date-time */
            readonly created_at: string;
            readonly aromatic_g: string;
            readonly diluent_g: string;
            readonly recipe_was_balanced: boolean;
            readonly days_macerating: number;
            /** Format: date */
            readonly ready_on: string;
            readonly days_remaining: number;
            readonly maceration_progress: string;
            readonly is_due: boolean;
            readonly ingredients: components["schemas"]["BatchIngredient"][];
            readonly notes: components["schemas"]["BatchNote"][];
        };
// ...
        PaginatedBatchList: {
            /** @example 123 */
            count: number;
            /**
             * Format: uri
             * @example http://api.example.org/accounts/?page=4
             */
            next?: string | null;
            /**
             * Format: uri
             * @example http://api.example.org/accounts/?page=2
             */
            previous?: string | null;
            results: components["schemas"]["Batch"][];
        };
// ...
        PaginatedRecipeList: {
            /** @example 123 */
            count: number;
            /**
             * Format: uri
             * @example http://api.example.org/accounts/?page=4
             */
            next?: string | null;
            /**
             * Format: uri
             * @example http://api.example.org/accounts/?page=2
             */
            previous?: string | null;
            results: components["schemas"]["Recipe"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}